requirements.txt
torch
torchvision
pillow
numpy
typing
scikit-image
huggingface_hub
transformers>=4.39.1
opencv-python
opencv-python-headless